requirements.txt
-i https://pypi.org/simple
anyio==3.6.2 ; python_full_version >= '3.6.2'
attrs==22.2.0 ; python_version >= '3.6'
boto3==1.26.65
botocore==1.29.65 ; python_version >= '3.7'
click==8.1.3 ; python_version >= '3.7'
contextlib2==21.6.0 ; python_version >= '3.6'
dill==0.3.6 ; python_version >= '3.7'
fastapi==0.89.1
google-pasta==0.2.0
h11==0.14.0 ; python_version >= '3.7'
httptools==0.5.0
idna==3.4 ; python_version >= '3.5'
importlib-metadata==4.13.0 ; python_version >= '3.7'
jmespath==1.0.1 ; python_version >= '3.7'
multiprocess==0.70.14 ; python_version >= '3.7'
numpy==1.24.2 ; python_version >= '3.8'
packaging==23.0 ; python_version >= '3.7'
pandas==1.5.3 ; python_version >= '3.8'
pathos==0.3.0 ; python_version >= '3.7'
pox==0.3.2 ; python_version >= '3.7'
ppft==1.7.6.6 ; python_version >= '3.7'
protobuf==3.20.3 ; python_version >= '3.7'
protobuf3-to-dict==0.1.5
pydantic==1.10.4 ; python_version >= '3.7'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.21.1
python-multipart==0.0.5
pytz==2022.7.1
pyyaml==6.0
s3transfer==0.6.0 ; python_version >= '3.7'
sagemaker==2.131.1
schema==0.7.5
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
smdebug-rulesconfig==1.0.1 ; python_version >= '2.7'
sniffio==1.3.0 ; python_version >= '3.7'
starlette==0.22.0 ; python_version >= '3.7'
typing-extensions==4.4.0 ; python_version >= '3.7'
urllib3==1.26.14 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
uvicorn[standard]==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
zipp==3.12.1 ; python_version >= '3.7'